Geometry class created
