+++ BusinessAssistant-opendata/src/main/java/com/businessassistantbcn/opendata/dto/largestablishments/ClassificationsDataDto.java
