+++ BusinessAssistant-opendata/src/main/java/com/businessassistantbcn/opendata/dto/largestablishments/CoordinateDto.java
